fontgan/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
brotlipy==0.7.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==37.0.1
cycler==0.11.0
Django==3.2.15
fonttools==4.37.3
idna==3.3
image==1.5.33
imageio==2.22.0
kiwisolver==1.4.4
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.21.6
packaging==21.3
Pillow==9.2.0
pip==22.1.2
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scipy==1.1.0
setuptools==63.4.1
six==1.16.0
sqlparse==0.4.2
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
typing_extensions==4.3.0
urllib3==1.26.11
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2

#pycopg2
#pydantic
#sqlalchemy
#opencv-python